requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django
django-allauth==0.49.0
django-appconf==1.0.5
django-crispy-forms==1.14.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
